temp 大致判定完成

+++ Assets/Script/NoR2252Application.cs
@@ -3,7 +3,7 @@ using NoR2252.Models;
 using UnityEngine;
 public static class NoR2252Application {
     static Sheet currentSheet = null;
-    static double videoTime = 0.0;
+    static float videoTime = 0f;
     public static Sheet CurrentSheet { get { return currentSheet; } set { currentSheet = value; } }
-    public static double VideoTime { get { return videoTime; } set { videoTime = value; } }
+    public static float VideoTime { get { return videoTime; } set { videoTime = value; } }
 }
